refactor: excluindo arquivos individuais de validação para unificar em um único arquivo.
