+++ src/modules/validacoes.h
@@ -1,0 +1,14 @@
+#ifndef VALIDACOES_H
+#define VALIDACOES_H
+
+int validaCPF(char *cpf);
+
+int validaTelefone(char *telefone);
+
+int validaNome(char *nome);
+
+int validaNascimento(char *nascimento);
+
+int validaEmail(char *email);
+
+#endif
